turn main into train
